Fusion_Portal/requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.22.0
Flask==3.0.3
Flask-Login==0.6.3
gunicorn==22.0.0
pyodbc==5.1.0

# Password hashing
passlib==1.7.4
bcrypt==4.0.1

python-dotenv==1.0.1
